requirements.txt
pytest-testconfig~=0.2.0
requests-toolbelt~=0.9.1
allure-pytest~=2.8.12
jsonschema~=3.2.0
termcolor~=1.1.0
requests~=2.23.0
colorlog~=4.1.0
pytruth~=1.1.0
pytest~=5.4.1
Faker~=4.0.2
pytest-xdist~=1.31.0
